Clean training/testing API
TripletBase is the basic triplet learning class.
It has a bunch of loss functions implemented along
with greedyTree generation. It has two unimplemented
functions:

1. nodeFeatures : dictionary of features for each node
2. embedding : consumes the output of nodeFeatures and
outputs embeddings.

Due to this, loss functions and data generation functions
can be written cleanly. Most of the other changes in this
commit are to facilitate this