Worked on refactoring and fixing bugs
1. All methods (baselines, our methods) now take a homogeneous input
that is an object of class SVGData and return an SVGData object too.
This helps in visualizing the output and reduces scope for making
mistakes such as forgetting to preprocess.
2. Deleted PatternGroupingV2 model since we haven't really used it.
3. Added containment guided tree inference to TripletBase
4. Removed options that we no longer care about. Like SWA.
5. Rewrote documentation for bipartite, boolean, boxes, datastructures
and comp modules. This revisit gave me a chance to catch bugs in unused
areas of the code.
6. Fixed a bug in parsing of CSS style strings. There are a lot
of corner cases in how these strings appear in the real world
that I wasn't aware about.
7. Significantly simplfied functions in treeOps.py