Added shape context based heuristic
The shape context heuristic is responsible for detecting
merges between a stroke and its fill. We check three things
while considering a pair for a merge.
  a. Whether its centroids are nearby. This property also
     speeds up the computation project since we can query
     nearby candidates fast using a KDTree.
  b. Whether the bounding box intersect above a threshold.
  c. Whether the shape contexts of the two shapes match.
Due to the use of KDTree and vectorized code, this is pretty
fast to compute and graphics having ~20 paths can be processed
within a second.